Remove decodeV2FundingCycleMetadata yay
