+++ src/components/v2/V2Project/V2PayButton/V2PayForm/V2PayForm.tsx
@@ -14,7 +14,7 @@ import { isAddress } from 'ethers/lib/utils'
 import { useContext, useState } from 'react'
 import {
   getUnsafeV2FundingCycleProperties,
-  V2FundingCycleRiskCount,
+  v2FundingCycleRiskCount,
 } from 'utils/v2/fundingCycle'
 
 import Sticker from 'components/icons/Sticker'
@@ -47,7 +47,8 @@ export const V2PayForm = ({
   const {
     theme: { colors },
   } = useContext(ThemeContext)
-  const { tokenAddress, fundingCycle } = useContext(V2ProjectContext)
+  const { tokenAddress, fundingCycle, fundingCycleMetadata } =
+    useContext(V2ProjectContext)
   const { projectMetadata } = useContext(ProjectMetadataContext)
 
   const [customBeneficiaryEnabled, setCustomBeneficiaryEnabled] =
@@ -63,9 +64,10 @@ export const V2PayForm = ({
   const canAddMoreStickers =
     (stickerUrls ?? []).length < ProjectPreferences.MAX_IMAGES_PAYMENT_MEMO
 
-  const riskCount = fundingCycle
-    ? V2FundingCycleRiskCount(fundingCycle)
-    : undefined
+  const riskCount =
+    fundingCycle && fundingCycleMetadata
+      ? v2FundingCycleRiskCount(fundingCycle, fundingCycleMetadata)
+      : undefined
 
   return (
     <>
@@ -263,9 +265,12 @@ export const V2PayForm = ({
         onCancel={() => setRiskModalVisible(false)}
         cancelText={<Trans>Close</Trans>}
       >
-        {fundingCycle && (
+        {fundingCycle && fundingCycleMetadata && (
           <ProjectRiskNotice
-            unsafeProperties={getUnsafeV2FundingCycleProperties(fundingCycle)}
+            unsafeProperties={getUnsafeV2FundingCycleProperties(
+              fundingCycle,
+              fundingCycleMetadata,
+            )}
           />
         )}
       </Modal>
