+++ src/components/v2/V2Project/V2ProjectSettings/pages/V2ProjectReconfigureFundingCycleSettingsPage/hooks/initialEditingData.ts
@@ -2,13 +2,9 @@ import useProjectDistributionLimit from 'hooks/v2/contractReader/ProjectDistribu
 import useProjectSplits from 'hooks/v2/contractReader/ProjectSplits'
 import { Split } from 'models/splits'
 import { useContext, useEffect, useState } from 'react'
-import {
-  defaultFundingCycleMetadata,
-  editingV2ProjectActions,
-} from 'redux/slices/editingV2Project'
+import { editingV2ProjectActions } from 'redux/slices/editingV2Project'
 import { fromWad } from 'utils/format/formatNumber'
 import { NO_CURRENCY, V2_CURRENCY_ETH } from 'utils/v2/currency'
-import { decodeV2FundingCycleMetadata } from 'utils/v2/fundingCycle'
 import {
   SerializedV2FundAccessConstraint,
   SerializedV2FundingCycleData,
@@ -59,6 +55,7 @@ export const useInitialEditingData = (visible: boolean) => {
   const {
     primaryTerminal,
     fundingCycle,
+    fundingCycleMetadata,
     payoutSplits,
     reservedTokensSplits,
     distributionLimit,
@@ -70,7 +67,8 @@ export const useInitialEditingData = (visible: boolean) => {
     projectId,
   })
 
-  const [queuedFundingCycle] = queuedFundingCycleResponse ?? []
+  const [queuedFundingCycle, queuedFundingCycleMetadata] =
+    queuedFundingCycleResponse ?? []
   const { data: queuedPayoutSplits } = useProjectSplits({
     projectId,
     splitGroup: ETH_PAYOUT_SPLIT_GROUP,
@@ -96,6 +94,10 @@ export const useInitialEditingData = (visible: boolean) => {
     ? queuedFundingCycle
     : fundingCycle
 
+  const effectiveFundingCycleMetadata = queuedFundingCycle?.number.gt(0)
+    ? queuedFundingCycleMetadata
+    : fundingCycleMetadata
+
   const effectivePayoutSplits = queuedFundingCycle?.number.gt(0)
     ? queuedPayoutSplits
     : payoutSplits
@@ -119,7 +121,8 @@ export const useInitialEditingData = (visible: boolean) => {
 
   // Populates the local redux state from V2ProjectContext values
   useEffect(() => {
-    if (!visible || !effectiveFundingCycle) return
+    if (!visible || !effectiveFundingCycle || !effectiveFundingCycleMetadata)
+      return
 
     // Build fundAccessConstraint
     let fundAccessConstraint: SerializedV2FundAccessConstraint | undefined =
@@ -161,21 +164,16 @@ export const useInitialEditingData = (visible: boolean) => {
       editingV2ProjectActions.setFundingCycleData(editingFundingCycleData),
     )
 
-    // Set editing funding metadata
-    const editingFundingCycleMetadata = effectiveFundingCycle.metadata
-      ? serializeV2FundingCycleMetadata(
-          decodeV2FundingCycleMetadata(effectiveFundingCycle.metadata),
-        )
-      : defaultFundingCycleMetadata
-    if (effectiveFundingCycle?.metadata) {
-      dispatch(
-        editingV2ProjectActions.setFundingCycleMetadata(
-          serializeV2FundingCycleMetadata(
-            decodeV2FundingCycleMetadata(effectiveFundingCycle.metadata),
-          ),
-        ),
-      )
-    }
+    const editingFundingCycleMetadata = serializeV2FundingCycleMetadata(
+      effectiveFundingCycleMetadata,
+    )
+
+    // Set editing funding cycle metadata
+    dispatch(
+      editingV2ProjectActions.setFundingCycleMetadata(
+        editingFundingCycleMetadata,
+      ),
+    )
 
     // Set editing payout splits
     dispatch(
@@ -201,6 +199,7 @@ export const useInitialEditingData = (visible: boolean) => {
   }, [
     contracts,
     effectiveFundingCycle,
+    effectiveFundingCycleMetadata,
     effectivePayoutSplits,
     effectiveReservedTokensSplits,
     effectiveDistributionLimit,
